# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: movie-api
  description: This is a REST API application made with Express. It serves a movies
    Database
  contact:
    name: movie-api
    url: https://github.com/Eloi-Perez/movie_api
  license:
    name: Licensed Under MIT
    url: https://spdx.org/licenses/MIT.html
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: Development server
paths:
...
  /img/{Title}:
    get:
      summary: Get Images by Title
      description: Movies images.
      operationId: imgTitleGET
      parameters:
      - name: Title
        in: path
        description: Title of the movie.
        required: true
        style: simple
        explode: false
        schema:
          type: array
          items: {}
      responses:
        "200":
          description: .jpg Movies images.
          content:
            image/jpeg: {}
        "404":
          description: Title not found.
      x-swagger-router-controller: Default
  /movies:
    get:
      summary: Get all Movies.
      description: Retrive the full list of movies and their properties
      operationId: moviesGET
      responses:
        "200":
          description: JSON array.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movies'
      security:
      - bearerAuth: []
      x-swagger-router-controller: Default
  /movies/featured:
    get:
      summary: Get Featured Movies.
      description: Retrive the featured list of movies and their properties
      operationId: moviesFeaturedGET
      responses:
        "200":
          description: JSON array.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movies'
      security:
      - bearerAuth: []
      x-swagger-router-controller: Default
  /movies/{Title}:
    get:
      summary: Get one Movie.
      description: Retrive movie {Title}.
      operationId: moviesTitleGET
      parameters:
      - name: Title
        in: path
        description: Title of the movie.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: JSON object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
      security:
      - bearerAuth: []
      x-swagger-router-controller: Default
  /genres/{Genre}:
    get:
      summary: Get genre info.
      description: Retrive genre {Genre}.
      operationId: genresGenreGET
      parameters:
      - name: Genre
        in: path
        description: Name of the genre.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: JSON object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
      security:
      - bearerAuth: []
      x-swagger-router-controller: Default
  /directors/{Director}:
    get:
      summary: Get director info.
      description: Retrive director {Director}.
      operationId: directorsDirectorGET
      parameters:
      - name: Director
        in: path
        description: Name of the director.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: JSON object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
      security:
      - bearerAuth: []
      x-swagger-router-controller: Default
...
components:
  schemas:
...
    Movie:
      type: object
      properties:
        _id:
          example: 60720485078f3662d0e67bfa
        Title:
          type: string
          example: Timecrimes
        Description:
          type: string
          example: A man accidentally gets into a time machine and...
        Genre:
          type: object
          example: {}
        Director:
          type: object
          example: {}
        Featured:
          type: boolean
          example: true
        ImagePath:
          type: string
          example: /img/timecrimes
      example:
        Description: A man accidentally gets into a time machine and...
        Director: {}
        ImagePath: /img/timecrimes
        Title: Timecrimes
        Featured: true
        _id: 60720485078f3662d0e67bfa
        Genre: {}
    Movies:
      type: array
      items:
        allOf:
        - $ref: '#/components/schemas/Movie'
    inline_response_200:
      type: object
      properties:
        _id:
          example: 60720485078f3662d0e67bfa
        Name:
          example: Drama
        Description:
          example: Focused on emotions and defined by conflict...
      example:
        Description: Focused on emotions and defined by conflict...
        _id: 60720485078f3662d0e67bfa
        Name: Drama
    inline_response_200_1:
      type: object
      properties:
        _id:
          example: 60720485078f3662d0e67bfa
        Name:
          example: Shane Carruth
        Description:
          example: is an American filmmaker, screenwriter, composer, and actor...
      example:
        Description: is an American filmmaker, screenwriter, composer, and actor...
        _id: 60720485078f3662d0e67bfa
        Name: Shane Carruth
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT